options: add new "discontig" option for engines
This option will enable capable engines to support sending discontiguous
IO buffers to the given device. Non zero value will hold the number of
maximal discontiguous elements in a single IO operation. For now, limit
this value to 4. This option will allow testing real world use cases
where the application can send discontiguous IO to the block device and
cause the block layer to use a bounce buffer. Some smart devices will be
able to handle discontiguous buffers and will not use the block layer to
fix the IO buffers alignment.

Also, block the verify option for discontiguous IO for now. This will be
enabled in the future.

Signed-off-by: Max Gurtovoy <mgurtovoy@nvidia.com>